requirements.txt
torchvision==0.2.2.post3 
torch==1.4.0
tensorflow-gpu==1.13.1
matplotlib==2.2.4
sklearn
scipy==1.0.1
scikit-image==0.14.2
opencv-python==4.1.0.25
easydict
tensorlayer==1.7
progressbar==2.5
numpy==1.15.1
tqdm
